+++ src/ServerDocumentor.php
@@ -134,14 +134,9 @@ class ServerDocumentor
 
 //'httpheaders' => array(),
 
-            'docheader' => '<?xml version="1.0" encoding="iso-8859-1"?>
-<!DOCTYPE html
-    PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN"
-    "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="{$lang}" lang="{$lang}">
+            'docheader' => '<!DOCTYPE html>
+<html lang="{$lang}">
 <head>
-<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1" />
-<meta http-equiv="Content-Script-Type" content="text-javascript" />
 <meta name="generator" content="' . PhpXmlRpc::$xmlrpcName . '" />
 <link rel="stylesheet" type="text/css" href="docxmlrpcs.css" />
 {$extras}