+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - bumped the minimum required version of the base phpxmlrpc library to 4.6
 - ported the Ajax module to contemporary php and jsxmlrpc. Replaced usage of javascript `alert` with `console.log`
+- updated the self-documenting server to generate html-5 in utf8 (was using xhtml, iso-8859-1 previously)
 - converted the remaining files to utf8
 - moved from Travis to GitHub Actions for testing; test on all php versions from 5.4 to 8.1
 - removed the WSDL schemas - they now reside at https://github.com/gggeek/xmlrpc-schemas and can be installed via Composer